requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
base58==2.1.1
bitarray==2.7.3
bitcoin==1.1.42
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
cryptography==40.0.2
cytoolz==0.12.1
Django==4.2.1
djangorestframework==3.14.0
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.0
exceptiongroup==1.1.1
frozenlist==1.3.3
hexbytes==0.3.0
idna==3.4
iniconfig==2.0.0
jsonschema==4.17.3
lru-dict==1.1.8
multidict==6.0.4
packaging==23.1
parsimonious==0.9.0
pluggy==1.0.0
protobuf==4.23.0
py-sr25519-bindings==0.2.0
pycparser==2.21
pycryptodome==3.17
pyOpenSSL==23.1.1
pyrsistent==0.19.3
pytest==7.3.1
pytz==2023.3
pywin32==306
regex==2023.5.5
requests==2.30.0
rlp==3.0.0
sqlparse==0.4.4
tomli==2.0.1
toolz==0.12.0
tzdata==2023.3
urllib3==2.0.2
web3==6.3.0
websockets==11.0.3
yarl==1.9.2
